Update file_disemvowel with Initial Argument Handling
In this commit, the main function is given a few cases to handle the possible combinations of arguments, where no input and output are given, where an input is given, but an output is not given; and where an input is not given, but an output is given. While the first five tests still pass, the last three as of now do not.

Co-authored-by: JadynSondrol <sondr073@morris.umn.edu>