Update summarize_tree.c with Logic for is_dir Function
In this commit, we add the basic logic that will be used by is_dur to determine if a given file is a directory or something else.

NOTE FOR LATER: If there is a substantial memory leak, check back here to see if there is a place where the memory must allocated.

Co-authored-by: JadynSondrol <sondr037@morris.umn.edu>