Update summarize_tree, summarize_ftw, and summarize_tree.sh to be Executable
In this commit, we update the file names of summarize_tree and summarize_ftw so that they are executable in bats. In addition, we add logic to summarize_tree.sh so that it can be executed as well. However, as of now, none of the test beyond existing and executing are passing.

Co-authored-by: JadynSondrol sondr037@morris.umn.edu